Tidy TableBuilder, refs 1750
