+++ src/SQLStore/TableBuilder/PostgresTableBuilder.php
@@ -8,7 +8,7 @@ namespace SMW\SQLStore\TableBuilder;
  *
  * @author mwjames
  */
-class PostgresRdbmsTableBuilder extends RdbmsTableBuilder {
+class PostgresTableBuilder extends TableBuilder {
 
 	/**
 	 * @since 2.5
@@ -46,24 +46,60 @@ class PostgresRdbmsTableBuilder extends RdbmsTableBuilder {
 		return false;
 	}
 
-	/**
-	 * @see RdbmsTableBuilder::getSQLFromDBName
-	 */
-	protected function getSQLFromDBName( array $tableOptions ) {
-		return '';
-	}
+	/** Create */
 
 	/**
-	 * @see RdbmsTableBuilder::getSQLFromDBTableOptions
+	 * @since 2.5
+	 *
+	 * {@inheritDoc}
 	 */
-	protected function getSQLFromDBTableOptions( array $tableOptions ) {
-		return '';
+	protected function doCreateTable( $tableName, array $tableOptions = null ) {
+
+		$tableName = $this->connection->tableName( $tableName );
+
+		$fieldSql = array();
+		$fields = $tableOptions['fields'];
+
+		foreach ( $fields as $fieldName => $fieldType ) {
+			$fieldSql[] = "$fieldName  $fieldType";
+		}
+
+		$this->connection->query( 'CREATE TABLE ' . $tableName . ' (' . implode( ',', $fieldSql ) . ') ', __METHOD__ );
 	}
 
+	/** Update */
+
 	/**
-	 * @see RdbmsTableBuilder::getCurrentFields
+	 * @since 2.5
+	 *
+	 * {@inheritDoc}
 	 */
-	protected function getCurrentFields( $tableName ) {
+	protected function doUpdateTable( $tableName, array $tableOptions = null ) {
+
+		$tableName = $this->connection->tableName( $tableName );
+		$currentFields = $this->getCurrentFields( $tableName );
+
+		$fields = $tableOptions['fields'];
+		$position = 'FIRST';
+
+		// Loop through all the field definitions, and handle each definition for either postgres or MySQL.
+		foreach ( $fields as $fieldName => $fieldType ) {
+			$this->doUpdateField( $tableName, $fieldName, $fieldType, $currentFields, $position, $tableOptions );
+
+			$position = "AFTER $fieldName";
+			$currentFields[$fieldName] = false;
+		}
+
+		// The updated fields have their value set to false, so if a field has a value
+		// that differs from false, it's an obsolete one that should be removed.
+		foreach ( $currentFields as $fieldName => $value ) {
+			if ( $value !== false ) {
+				$this->doDropField( $tableName, $fieldName );
+			}
+		}
+	}
+
+	private function getCurrentFields( $tableName ) {
 
 		$tableName = str_replace( '"', '', $tableName );
 		// Use the data dictionary in postgresql to get an output comparable to DESCRIBE.
@@ -122,17 +158,7 @@ EOT;
 		return $currentFields;
 	}
 
-	/**
-	 * @see RdbmsTableBuilder::doDropTable
-	 */
-	protected function doDropTable( $tableName ) {
-		$this->connection->query( 'DROP TABLE IF EXISTS ' . $this->connection->tableName( $tableName ), __METHOD__ );
-	}
-
-	/**
-	 * @see RdbmsTableBuilder::doUpdateField
-	 */
-	protected function doUpdateField( $tableName, $fieldName, $fieldType, $currentFields, $position, array $tableOptions ) {
+	private function doUpdateField( $tableName, $fieldName, $fieldType, $currentFields, $position, array $tableOptions ) {
 
 		$keypos = strpos( $fieldType, ' PRIMARY KEY' );
 
@@ -179,17 +205,43 @@ EOT;
 		}
 	}
 
-	/**
-	 * @see RdbmsTableBuilder::doDropField
-	 */
-	protected function doDropField( $tableName, $fieldName ) {
+	private function doDropField( $tableName, $fieldName ) {
+		$this->reportMessage( "   ... deleting obsolete field $fieldName ... " );
 		$this->connection->query( 'ALTER TABLE ' . $tableName . ' DROP COLUMN "' . $fieldName . '"', __METHOD__ );
+		$this->reportMessage( "done.\n" );
 	}
 
+	/** Index */
+
 	/**
-	 * @see RdbmsTableBuilder::doDropObsoleteIndicies
+	 * @since 2.5
+	 *
+	 * {@inheritDoc}
 	 */
-	protected function doDropObsoleteIndicies( $tableName, array &$indicies ) {
+	protected function doCreateIndicies( $tableName, array $indexOptions = null ) {
+
+		$indicies = $indexOptions['indicies'];
+
+		// First remove possible obsolete indicies
+		$this->doDropObsoleteIndicies( $tableName, $indicies );
+
+		// Add new indexes.
+		foreach ( $indicies as $indexName => $index ) {
+			// If the index is an array, it contains the column
+			// name as first element, and index type as second one.
+			if ( is_array( $index ) ) {
+				$columns = $index[0];
+				$indexType = count( $index ) > 1 ? $index[1] : 'INDEX';
+			} else {
+				$columns = $index;
+				$indexType = 'INDEX';
+			}
+
+			$this->doCreateIndex( $tableName, $indexType, $indexName, $columns, $indexOptions );
+		}
+	}
+
+	private function doDropObsoleteIndicies( $tableName, array &$indicies ) {
 
 		$tableName = $this->connection->tableName( $tableName, 'raw' );
 		$currentIndicies = $this->getIndexInfo( $tableName );
@@ -209,10 +261,7 @@ EOT;
 		}
 	}
 
-	/**
-	 * @see RdbmsTableBuilder::doCreateIndex
-	 */
-	protected function doCreateIndex( $tableName, $indexType, $indexName, $columns, array $indexOptions ) {
+	private function doCreateIndex( $tableName, $indexType, $indexName, $columns, array $indexOptions ) {
 
 		if ( $indexType === 'FULLTEXT' ) {
 			return $this->reportMessage( "   ... skipping the fulltext index creation ..." );
@@ -263,6 +312,17 @@ EOT;
 		$this->reportMessage( "   ... removing index $columns ..." );
 		$this->connection->query( 'DROP INDEX IF EXISTS ' . $indexName, __METHOD__ );
 		$this->reportMessage( "done.\n" );
+	}
+
+	/** Drop */
+
+	/**
+	 * @since 2.5
+	 *
+	 * {@inheritDoc}
+	 */
+	protected function doDropTable( $tableName ) {
+		$this->connection->query( 'DROP TABLE IF EXISTS ' . $this->connection->tableName( $tableName ), __METHOD__ );
 	}
 
 }
